Added tinyag to LengthOfAudio
I found quite an interesting conundrum, Wav files are often used to hold important metadata that otherwise often gets lost with other file formats when the files have been modified. The artist tag is used to store important JSON Metadata in this example. However, Tiny tag can't read my wav unless its mp3. In the properties it displays the data for wav, and not so for mp3. Interesting.